requirements.txt
torch
torchvision
flask
pytorch-msssim
Pillow
